requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-Cors==3.0.10
Flask-WTF==0.15.1
python-dotenv==0.19.0
Werkzeug==2.0.1
MarkupSafe==2.0.1
Eventlet==0.33.0
python-engineio==4.3.0
python-socketio==5.5.0
SQLAlchemy==1.4.23
requests==2.26.0
uuid==1.30 